`Task:` Adding `TypeScript` & moving into .src
