--- components/icons/gotpop.js
@@ -1,13 +1,0 @@
-import Link from "next/link";
-
-export default function IconGotPop(props) {
-    return (
-        <Link href="/">
-            <a>
-                <svg width="301.344" height="73.189" viewBox="0 0 79.424 19.364" className="max-w-[180px]" name="gotpop">
-                    <path className="logo-path animate-logoStartUp animation-delay-200 fill-zinc-600 dark:fill-white" d="M64.852 18.85l.198-.782a5307.07 5307.07 0 0 0 3.066-12.561l.108-.426h1.77c1.623 0 1.761 0 1.761.03s-.069.376-.178.781a9.762 9.762 0 0 0-.168.732.692.692 0 0 0 .149-.138 4.629 4.629 0 0 1 2.571-1.563c.297-.05.93-.06 1.266 0 .732.1 1.415.435 1.84.9.485.515.781 1.217.92 2.127a8.259 8.259 0 0 1-.96 4.965 5.252 5.252 0 0 1-3.481 2.75c-.247.049-.356.059-.85.049a3.125 3.125 0 0 1-2.315-.83 3.224 3.224 0 0 1-.742-1.237c-.089-.257-.08-.277-.198.277a845.895 845.895 0 0 1-1.206 4.906c-.01.04-.12.04-1.78.04h-1.781zm7.329-5.065c.296-.099.494-.227.751-.494.713-.722 1.227-1.949 1.494-3.59.07-.396.07-.495.07-1.049 0-.554 0-.623-.05-.811-.198-.722-.653-1.088-1.345-1.088-.851 0-1.583.752-2.176 2.196a9 9 0 0 0-.584 2.74c-.06 1.414.495 2.235 1.474 2.175.099-.01.267-.04.366-.079zM7.1 15.714a7.16 7.16 0 0 1-4.688-2.047A6.39 6.39 0 0 1 .75 10.145a8.546 8.546 0 0 1 0-2.057 8.11 8.11 0 0 1 1.919-4.193A9.347 9.347 0 0 1 8.76.739c.416-.04 1.385-.04 1.76 0 1.277.12 2.394.554 3.285 1.237.227.178.662.593.85.82a5.43 5.43 0 0 1 1.138 2.74c.04.337.06 1 .04 1.177l-.02.149h-1.118l-1.128-.02v-.445a4.985 4.985 0 0 0-.237-1.543 3.145 3.145 0 0 0-2.522-2.245 5.667 5.667 0 0 0-1.177-.02c-2.166.208-3.897 1.968-4.54 4.619a6.884 6.884 0 0 0-.228 2.087c0 .85.03 1.226.178 1.8.396 1.622 1.316 2.621 2.671 2.928.84.197 2.028.04 2.78-.337.098-.059.098-.03.227-.593a312.003 312.003 0 0 1 .702-2.987c0-.02-.326-.03-1.236-.03H8.959l.05-.217a1080.889 1080.889 0 0 0 .415-1.642c.03-.03.683-.03 3.126-.03h3.095v.05l-.03.148-.197.781a2378.523 2378.523 0 0 1-1.484 6.152l-.109.416h-1.483l-.396-.267-.623-.436-.218-.148-.237.119c-.722.346-1.286.534-2.037.653a16.38 16.38 0 0 1-1.731.059zm14.589 0a5.064 5.064 0 0 1-4.471-4.876c0-.534.03-.801.148-1.286a6.429 6.429 0 0 1 4.64-4.48 5.41 5.41 0 0 1 3.876.365 5.173 5.173 0 0 1 2.048 1.91 5.62 5.62 0 0 1 .573 1.582c.08.445.08 1.187 0 1.642a5.885 5.885 0 0 1-1.651 3.145 6.508 6.508 0 0 1-5.163 1.998zm.761-1.474c.297-.069.495-.198.791-.494.396-.386.693-.85.99-1.543.365-.89.623-1.88.78-2.997.05-.336.06-.554.06-1.019 0-.544 0-.623-.06-.791-.197-.722-.632-1.049-1.354-1.058-.327 0-.544.069-.821.247-.87.593-1.622 2.196-1.978 4.282a7.715 7.715 0 0 0-.08 2.325c.14.583.505.97 1.05 1.068.197.03.444.02.622-.02zm10.603 1.484a6.528 6.528 0 0 1-.791-.178 2.562 2.562 0 0 1-1.82-2.156c-.05-.297-.02-.97.05-1.276a582.712 582.712 0 0 1 1.325-5.341l-.742-.01c-.692 0-.742 0-.742-.05l.198-.84.198-.782.761-.01h.762l.89-.87a815.461 815.461 0 0 0 1.75-1.711l.15-.129h.75c.406 0 .743 0 .743.02a93.645 93.645 0 0 1-.633 2.64c0 .04.158.04 1.473.04 1.019 0 1.484.01 1.484.03l-.396 1.563-.02.099h-2.996l-.752 3.007a252.669 252.669 0 0 0-.781 3.214c-.06.267-.05.514.02.643.118.188.434.267.711.188a.663.663 0 0 0 .317-.198c.297-.326.534-.89.98-2.403l.058-.198h.782c.435 0 .791 0 .791.02a11.038 11.038 0 0 1-.83 2.522c-.515 1.068-1.158 1.73-1.95 2.008-.355.118-.553.148-1.117.158h-.633zm24.588-.05a5.044 5.044 0 0 1-4.124-3.936 5.845 5.845 0 0 1-.05-1.434 6.132 6.132 0 0 1 3.353-4.688 6.736 6.736 0 0 1 2.226-.693 5.104 5.104 0 0 1 5.005 2.235 4.916 4.916 0 0 1 .366 4.55 6.637 6.637 0 0 1-5.866 4.016c-.524 0-.613 0-.91-.05zm.99-1.424c.998-.178 1.928-1.711 2.442-4.025.179-.841.297-1.93.228-2.434a2.047 2.047 0 0 0-.198-.761c-.198-.396-.534-.633-1.019-.692-.534-.05-.93.118-1.365.544-.761.78-1.404 2.423-1.661 4.292a9.624 9.624 0 0 0-.099 1.583c.04.494.178.91.425 1.167.297.296.792.415 1.246.326zm-19.624 1.266l.208-.9A51352.116 51352.116 0 0 1 42.617.947l.02-.06h3.205c2.809.01 3.254.02 3.58.06 1.494.168 2.611.623 3.373 1.355.524.505.84 1.108.94 1.82.059.366.059.593 0 .92a3.808 3.808 0 0 1-1.385 2.403 7.715 7.715 0 0 1-3.264 1.484c-.97.198-1.88.277-3.62.277h-1.147l-.02.079-.08.297a2711.88 2711.88 0 0 0-1.127 4.55l-.336 1.404h-1.88c-1.463 0-1.879 0-1.879-.03zm8.041-8.229c1.177-.148 2.117-.742 2.483-1.582.158-.347.198-.505.198-.92.03-.436-.01-.663-.168-.97a1.681 1.681 0 0 0-.732-.702c-.524-.257-.792-.297-1.959-.316l-.94-.01-.237.94-.88 3.56c0 .04.05.04.98.04.692 0 1.048 0 1.265-.04z" stroke="#000" strokeWidth="0.989"></path>
-                </svg>
-            </a>
-        </Link>
-    );
-}
