--- jsconfig.json
@@ -1,19 +1,0 @@
-{
-    "compilerOptions": {
-      "baseUrl": ".",
-      "paths": {
-        "@lib/*": [
-          "./lib/*"
-        ],
-        "@utils/*": [
-          "./utils/*"
-        ],
-        "@components/*": [
-          "./components/*"
-        ],
-        "@studio/*": [
-          "./studio/*"
-        ]
-      }
-    }
-  }
