+++ next.config.js
@@ -1,11 +1,5 @@
 module.exports = {
   images: {
     domains: ["cdn.sanity.io", "avatars.githubusercontent.com"],
-    // loader: "custom"
-  },
-  swcMinify: true,
-  experimental: {
-    legacyBrowsers: false,
-    browsersListForSwc: true
   }
 };
