+++ package.json
@@ -1,5 +1,5 @@
 {
-  "private": true,
+  "private": false,
   "scripts": {
     "dev": "next dev",
     "prebuild": "echo 'Building Sanity to public/studio' && cd studio && yarn && npx @sanity/cli build ../public/studio -y && echo 'Done'",
@@ -15,12 +15,12 @@
     "@heroicons/react": "^1.0.6",
     "@portabletext/react": "^1.0.6",
     "@sanity/block-content-to-react": "^3.0.0",
-    "@sanity/core": "^2.33.0",
+    "@sanity/core": "^2.35.6",
     "@sanity/image-url": "^1.0.1",
     "@tailwindcss/line-clamp": "^0.4.0",
     "@tailwindcss/typography": "^0.5.2",
     "date-fns": "^2.28.0",
-    "next": "^12.1.6",
+    "next": "^13.1.6",
     "next-sanity": "^0.5.2",
     "next-sanity-image": "^3.2.1",
     "next-seo": "^5.4.0",
@@ -29,14 +29,17 @@
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.31.2",
     "tailwindcss-animation-delay": "^1.0.7",
+    "ts-node": "^10.9.1",
     "use-web3forms": "^1.2.1"
   },
   "devDependencies": {
+    "@types/node": "^18.13.0",
     "autoprefixer": "^10.4.7",
     "eslint": "^8.14.0",
     "eslint-config-next": "^12.1.6",
     "eslint-config-sanity": "^6.0.0",
     "postcss": "^8.4.13",
-    "tailwindcss": "^3.0.24"
+    "tailwindcss": "^3.0.24",
+    "typescript": "^4.9.5"
   }
 }
