+++ src/components/footer.js
@@ -1,6 +1,6 @@
 import Container from "@components/container";
-import ThemeSwitch from "@components/themeSwitch";
 import Social from "@components/social";
+import ThemeSwitch from "@components/themeSwitch";
 
 export default function Footer(props) {
   return (
