+++ src/components/layout.js
@@ -1,7 +1,7 @@
-import React from "react";
+import Footer from "@components/footer";
 import Head from "next/head";
 import Navbar from "@components/navbar";
-import Footer from "@components/footer";
+import React from "react";
 
 export default function Layout(props) {
   const { children } = props;
