+++ src/components/pagination.js
@@ -1,9 +1,10 @@
+import { ArrowCircleLeftIcon, ArrowCircleRightIcon } from '@heroicons/react/outline'
+import { calculatePrevNext, handleKeyDown } from '@utils/pagination'
+import { useEffect, useState } from 'react'
+
 import Container from '@components/container'
 import Link from 'next/link'
-import { ArrowCircleLeftIcon, ArrowCircleRightIcon } from '@heroicons/react/outline'
-import { useEffect, useState } from 'react'
 import { useRouter } from 'next/router'
-import { calculatePrevNext, handleKeyDown } from '@utils/pagination'
 
 export default function Pagination({ projects, post }) {
   const [next, setNext] = useState()
@@ -35,17 +36,13 @@ export default function Pagination({ projects, post }) {
     <Container>
       <article className="max-w-screen-md mx-auto ">
         <div className="flex justify-center mt-7 mb-7">
-          <Link href={`/project/${previous}`}>
-            <a className="flex items-center px-4 py-2 text-sm text-blue-600 rounded-full dark:text-blue-500 bg-brand-secondary/20">
+          <Link className="flex items-center px-4 py-2 text-sm text-blue-600 rounded-full dark:text-blue-500 bg-brand-secondary/20" href={`/project/${previous}`}>
               <ArrowCircleLeftIcon className="mr-2 w-7 h-7" />
               <span className="uppercase">Prev</span>
-            </a>
           </Link>
-          <Link href={`/project/${next}`}>
-            <a className="flex items-center px-4 py-2 text-sm text-blue-600 rounded-full dark:text-blue-500 bg-brand-secondary/20">
+          <Link className="flex items-center px-4 py-2 text-sm text-blue-600 rounded-full dark:text-blue-500 bg-brand-secondary/20" href={`/project/${next}`}>
               <span className="uppercase">Next</span>
               <ArrowCircleRightIcon className="ml-2 w-7 h-7" />
-            </a>
           </Link>
         </div>
       </article>
