+++ src/pages/skills/index.js
@@ -1,23 +1,19 @@
-import Head from 'next/head';
-import Layout from "@components/layout";
-import Container from "@components/container";
-import ProjectList from "@components/projectlist";
+import { allcatquery, configQuery, postquery } from "@lib/groq";
+import { getClient, usePreviewSubscription } from "@lib/sanity";
+
 import CategoryLabel from "@components/projects/category";
+import Container from "@components/container";
 import Intro from "@components/ui/intro";
+import Layout from "@components/layout";
+import ProjectList from "@components/projectlist";
 import { useRouter } from "next/router";
-import client, { getClient, usePreviewSubscription } from "@lib/sanity";
-import { configQuery, pathquery, catquery, allcatquery } from "@lib/groq";
-import { useEffect, useState } from 'react';
 
-export default function Post(props) {
-  const { siteconfig, preview, projectdata, categorydata, profile } = props;
+export default function Skills(props) {
+  const { postdata, siteconfig, preview, categorydata, profile } = props;
   const router = useRouter();
-  const { slug } = router.query;
-  const [title, setTitle] = useState();
 
-  const { data: projects } = usePreviewSubscription(catquery, {
-    params: { slug: slug },
-    initialData: projectdata,
+  const { data: projects } = usePreviewSubscription(postquery, {
+    initialData: postdata,
     enabled: preview || router.query.preview !== undefined
   });
 
@@ -31,27 +27,16 @@ export default function Post(props) {
     enabled: preview || router.query.preview !== undefined
   });
 
-  useEffect(() => {
-    categories?.forEach(category => {
-      if (category.slug.current !== slug) return
-
-      setTitle(category.title)
-    });
-  }, [categories, slug])
-
   return (
     <>
-      <Head>
-        <title>Skills</title>
-      </Head>
       {projects && siteConfig && (
         <Layout {...siteConfig}>
           <Container>
             <Intro
-              title={title}
+              title={'Skills'}
               profile={profile}
-              skills={true}
-              left={<CategoryLabel categories={categories} />} />
+              left={<CategoryLabel categories={categories} />}
+            />
             <div className="grid gap-10 mt-10 lg:gap-10 md:grid-cols-2 xl:grid-cols-3 ">
               {projects.map(project => (
                 <ProjectList
@@ -69,11 +54,9 @@ export default function Post(props) {
 }
 
 export async function getStaticProps({ params, preview = false }) {
+  const post = await getClient(preview).fetch(postquery);
   const config = await getClient(preview).fetch(configQuery);
   const categoriesList = await getClient(preview).fetch(allcatquery);
-  const projects = await getClient(preview).fetch(catquery, {
-    slug: params.slug
-  });
 
   const url = "https://api.github.com/users/gotpop"
   const getProfile = await fetch(url)
@@ -81,7 +64,7 @@ export async function getStaticProps({ params, preview = false }) {
 
   return {
     props: {
-      projectdata: projects,
+      postdata: post,
       profile: getProfileData,
       categorydata: categoriesList,
       siteconfig: { ...config },
@@ -89,18 +72,4 @@ export async function getStaticProps({ params, preview = false }) {
     },
     revalidate: 10
   };
-}
-
-export async function getStaticPaths({ params }) {
-  const allPosts = await client.fetch(pathquery);
-
-  return {
-    paths:
-      allPosts?.map(page => ({
-        params: {
-          slug: page.slug.current
-        }
-      })) || [],
-    fallback: true
-  }
 }
