+++ src/utils/upperFirst.js
